finsh GetCart in CartApiController in ShoppingCartApi

+++ Mango.Services.ShoppingCartApi/Models/Dto/CartHeaderDto.cs
@@ -9,7 +9,7 @@ namespace Mango.Services.ShoppingCartApi.Models.Dto
         public int CartHeaderId { get; set; }
         public string? UserId { get; set; }
         public string? CouponCode { get; set; }
-        public decimal Discount { get; set; }
-        public decimal CartTotal { get; set; }
+        public double Discount { get; set; }
+        public double CartTotal { get; set; }
     }
 }
